refactor: add hook useFetchMovies
