+++ src/components/MovieSection/index.tsx
@@ -4,30 +4,10 @@ import InputText from "../InputText";
 import Button from "../Button";
 import { FaSearch } from "react-icons/fa";
 import MovieList from "../MovieList";
-import { useEffect, useState } from "react";
-import type { Movie } from "../../Types";
-import { getMovies } from "../../Api";
+import useFetchMovies from "../../Hooks/useFetchMovies";
 
 const MovieSection = () => {
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [error, setError] = useState<string | null>(null);
-
-  const fetchMovies = async () => {
-    try {
-      const movies = await getMovies();
-      setMovies(movies);
-      setError(null);
-    } catch {
-      setError(
-        "Não foi possível carregar os filmes. Tente novamente mais tarde."
-      );
-    }
-  };
-
-  useEffect(() => {
-    fetchMovies();
-  }, []);
-
+  const { movies, error, isLoading } = useFetchMovies();
   return (
     <main>
       <section className={styles.container}>
@@ -39,7 +19,11 @@ const MovieSection = () => {
         </Fieldset>
         <h1 className={styles.title}>Em cartaz</h1>
         {error && <div className={styles.error}>{error}</div>}
-        <MovieList movies={movies}></MovieList>
+        {isLoading ? (
+          <div className={styles.loading}>Carregando...</div>
+        ) : (
+          <MovieList movies={movies} />
+        )}
       </section>
     </main>
   );
